app/RegistrationLoginValidation/LoginService/requirements.txt
anyio==3.6.2
astroid==2.15.0
click==8.1.3
coverage==7.2.2
dill==0.3.6
fastapi==0.95.0
Flask==2.2.3
h11==0.14.0
idna==3.4
importlib-metadata==6.3.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
platformdirs==3.1.1
psycopg2-binary==2.9.6
pydantic==1.10.7
pylint==2.17.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
utils==1.0.1
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
requests